Extract location
